+++ lib/pages/settings/restore_db_page.dart
@@ -1,0 +1,14 @@
+import 'dart:io';
+import 'package:flutter/cupertino.dart';
+import 'package:flutter/material.dart';
+import 'package:flutter/rendering.dart';
+import 'package:flutter/services.dart';
+import 'package:npng/generated/l10n.dart';
+import 'package:npng/widgets/bottom_bar.dart';
+import 'package:npng/widgets/multiplatform_widgets.dart';
+import 'package:package_info_plus/package_info_plus.dart';
+import 'package:page_transition/page_transition.dart';
+import 'package:flash/flash.dart';
+import 'package:npng/db.dart';
+import 'about_page.dart';
+import 'package:file_picker/file_picker.dart';