+++ CHANGELOG.md
@@ -2,6 +2,11 @@
 - Changed alarm sound.
 - Added "Stop the rest" button on timer screen.
 - Resetting workout data after finish.
+- New Settings screen.
+- Share database.
+- Backup database.
+- Restore database.
+- Import database.
 
 ## 1.0
 First version.